:hammer: Add ability to add item to card on details page

+++ t-shirt/src/components/ItemDetails.js
@@ -9,6 +9,7 @@ function ItemDetails({ match }) {
     const [deleted, setDeleted] = useState(false);
     const [item, setItem] = useState({});
     const [editing, setEditing] = useState(false);
+    const [showCart, setShowCart] = useState(false);
 
     useEffect(() => {
         axios
@@ -21,7 +22,40 @@ function ItemDetails({ match }) {
             });
     }, [match.params.id]);
 
-    const deleteItem = (event) => {
+    const addItemToCart = () => {
+        axios
+            .get(`${APIURL}/order`)
+            .then((response) => {
+                const order = response.data[0];
+                const checkIndex = order.items.findIndex(
+                    (element) => element._id === item._id
+                );
+
+                if (checkIndex === -1) {
+                    const updatedOrder = {
+                        ...order,
+                        items: [...order.items, item._id],
+                    };
+                    console.log(updatedOrder);
+                    return updatedOrder;
+                }
+                return order;
+            })
+            .then((orderUpdate) => {
+                return axios.put(
+                    `${APIURL}/order/${orderUpdate._id}`,
+                    orderUpdate
+                );
+            })
+            .then(() => {
+                setShowCart(true);
+            })
+            .catch(function (error) {
+                console.log(error);
+            });
+    };
+
+    const deleteItem = () => {
         const url = `${APIURL}/items/${match.params.id}`;
 
         axios
@@ -44,17 +78,24 @@ function ItemDetails({ match }) {
         return <Redirect to={`/items/${match.params.id}/edit`} />;
     }
 
+    if (showCart) {
+        return <Redirect to="/order" />;
+    }
+
     return (
         <div className="item-details-container">
             <h2> {item.name} </h2>
             <h3> {item.description} </h3>
             <h3> {item.price} </h3>
 
-            <Button variant="secondary" onClick={deleteItem}>
-                Buy
+            <Button variant="secondary" onClick={addItemToCart}>
+                Add To Cart
             </Button>
             <Button variant="secondary" onClick={editItem}>
                 Edit
+            </Button>
+            <Button variant="secondary" onClick={deleteItem}>
+                Delete
             </Button>
         </div>
     );
